fix: amount 데이터타입 Integer -> Long

+++ src/main/java/christmas/view/OutputView.java
@@ -21,7 +21,7 @@ public class OutputView {
         }
     }
 
-    public void printTotalAmount(long amount) {
+    public void printTotalAmount(Long amount) {
         System.out.println("할인 전 총주문 금액");
         System.out.println(amount);
     }
@@ -44,19 +44,19 @@ public class OutputView {
         System.out.println("없음");
     }
 
-    public void printD_DayDiscount(Integer amount) {
+    public void printD_DayDiscount(Long amount) {
         System.out.println("크리스마스 디데이 할인: -" + amount + "원");
     }
 
-    public void printWeekdayDiscount(Integer amount) {
+    public void printWeekdayDiscount(Long amount) {
         System.out.println("평일 할인: -" + amount + "원");
     }
 
-    public void printWeekendDiscount(Integer amount) {
+    public void printWeekendDiscount(Long amount) {
         System.out.println("주말 할인: -" + amount + "원");
     }
 
-    public void printSpecialDiscount(Integer amount) {
+    public void printSpecialDiscount(Long amount) {
         System.out.println("특별 할인:: -" + amount + "원");
     }
 
@@ -64,12 +64,12 @@ public class OutputView {
         System.out.println("증정 이벤트: -" + gift.getPrice() + "원");
     }
 
-    public void printTotalBenefitAmount(Integer amount) {
+    public void printTotalBenefitAmount(Long amount) {
         System.out.println("<총혜택 금액>");
         System.out.println("-" + amount + "원");
     }
 
-    public void printAfterDiscountAmount(Integer amount) {
+    public void printAfterDiscountAmount(Long amount) {
         System.out.println("<할인 후 예상 결제 금액>");
         System.out.println(amount + "원");
     }
